provide a fillDescriptions for {Global}DetLayerGeometryESProducer
