+++ TrackingTools/RecoGeometry/plugins/DetLayerGeometryESProducer.cc
@@ -1,12 +1,23 @@
-#include "TrackingTools/RecoGeometry/plugins/DetLayerGeometryESProducer.h"
-
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/Framework/interface/ESHandle.h"
-#include "FWCore/Framework/interface/ModuleFactory.h"
-#include "FWCore/Framework/interface/ESProducer.h"
-
 #include <memory>
 #include <string>
+
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "TrackingTools/DetLayers/interface/DetLayerGeometry.h"
+#include "TrackingTools/RecoGeometry/interface/RecoGeometryRecord.h"
+
+class DetLayerGeometryESProducer : public edm::ESProducer {
+public:
+  DetLayerGeometryESProducer(const edm::ParameterSet& p);
+  ~DetLayerGeometryESProducer() override = default;
+
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+  std::unique_ptr<DetLayerGeometry> produce(const RecoGeometryRecord&);
+};
 
 using namespace edm;
 
@@ -15,7 +26,11 @@ DetLayerGeometryESProducer::DetLayerGeometryESProducer(const edm::ParameterSet& 
   setWhatProduced(this, myName);
 }
 
-DetLayerGeometryESProducer::~DetLayerGeometryESProducer() {}
+void DetLayerGeometryESProducer::fillDescriptions(edm::ConfigurationDescriptions& descriptions) {
+  edm::ParameterSetDescription desc;
+  desc.add<std::string>("ComponentName", std::string(""));
+  descriptions.addWithDefaultLabel(desc);
+}
 
 std::unique_ptr<DetLayerGeometry> DetLayerGeometryESProducer::produce(const RecoGeometryRecord& iRecord) {
   return std::make_unique<DetLayerGeometry>();
